// components-pro-display/button/enum.tsx
export enum ButtonColor {
  default = 'default',
  primary = 'primary',
  secondary = 'secondary',
  gray = 'gray',
  blue = 'blue',
  red = 'red',
  green = 'green',
  yellow = 'yellow',
  purple = 'purple',
  dark = 'dark',
}

export enum FuncType {
  flat = 'flat',
  raised = 'raised',
  link = 'link',
}

export enum ButtonType {
  button = 'button',
  submit = 'submit',
  reset = 'reset',
}

export enum ButtonTooltip {
  none = 'none',
  always = 'always',
  overflow = 'overflow',
}

export enum WaitType {
  throttle = 'throttle',
  debounce = 'debounce',
}